+++ scripts/backup-tool.py
@@ -1,0 +1,7 @@
+#! /usr/bin/python
+"""Create a backup.
+"""
+
+import archive.bt
+
+archive.bt.backup_tool()